Split contracts over good and bad
