+++ plugins/formal-verification/testData/diagnostics/bad_contracts/calls_in_place.fir.diag.txt
@@ -1,0 +1,101 @@
+/calls_in_place.kt:(198,205): info: Generated Viper text for unknown:
+method pkg_$global$unknown(local$f: Ref) returns (ret: Int)
+  requires dom$Type$isSubtype((dom$TypeOf$typeOf(local$f): dom$Type), dom$Type$Function())
+  requires acc(local$f.special$function_object_call_counter, write)
+  ensures acc(local$f.special$function_object_call_counter, write)
+  ensures old(local$f.special$function_object_call_counter) <=
+    local$f.special$function_object_call_counter
+  ensures true
+{
+  var anonymous$1: Int
+  special$invoke_function_object(local$f)
+  ret := anonymous$1
+}
+
+/calls_in_place.kt:(345,367): info: Generated Viper text for incorrect_at_most_once:
+method pkg_$global$unknown(local$f: Ref) returns (ret: Int)
+  requires dom$Type$isSubtype((dom$TypeOf$typeOf(local$f): dom$Type), dom$Type$Function())
+  requires acc(local$f.special$function_object_call_counter, write)
+  ensures acc(local$f.special$function_object_call_counter, write)
+  ensures old(local$f.special$function_object_call_counter) <=
+    local$f.special$function_object_call_counter
+  ensures true
+
+
+method pkg_$global$incorrect_at_most_once(local$f: Ref) returns (ret: Int)
+  requires dom$Type$isSubtype((dom$TypeOf$typeOf(local$f): dom$Type), dom$Type$Function())
+  requires acc(local$f.special$function_object_call_counter, write)
+  ensures acc(local$f.special$function_object_call_counter, write)
+  ensures old(local$f.special$function_object_call_counter) <=
+    local$f.special$function_object_call_counter
+  ensures local$f.special$function_object_call_counter <=
+    old(local$f.special$function_object_call_counter) + 1
+{
+  var anonymous$1: Int
+  var anonymous$2: Int
+  anonymous$1 := pkg_$global$unknown(local$f)
+  special$invoke_function_object(local$f)
+  ret := anonymous$2
+}
+
+/calls_in_place.kt:(345,367): warning: Viper verification error: Postcondition of pkg_$global$incorrect_at_most_once might not hold. Assertion local$f.special$function_object_call_counter <= old(local$f.special$function_object_call_counter) + 1 might not hold. (<no position>)
+
+
+/calls_in_place.kt:(345,367): warning: Function incorrect_at_most_once may not satisfy its contract.
+
+/calls_in_place.kt:(397,451): warning: Wrong invocation kind 'AT_MOST_ONCE' for 'f: (Int) -> Int' specified, the actual invocation kind is 'AT_LEAST_ONCE'.
+
+/calls_in_place.kt:(521,543): info: Generated Viper text for incorrect_exactly_once:
+method pkg_$global$incorrect_exactly_once(local$f: Ref) returns (ret: Int)
+  requires dom$Type$isSubtype((dom$TypeOf$typeOf(local$f): dom$Type), dom$Type$Function())
+  requires acc(local$f.special$function_object_call_counter, write)
+  ensures acc(local$f.special$function_object_call_counter, write)
+  ensures old(local$f.special$function_object_call_counter) <=
+    local$f.special$function_object_call_counter
+  ensures local$f.special$function_object_call_counter ==
+    old(local$f.special$function_object_call_counter) + 1
+{
+  var anonymous$1: Int
+  var anonymous$2: Int
+  special$invoke_function_object(local$f)
+  special$invoke_function_object(local$f)
+  ret := anonymous$2
+}
+
+/calls_in_place.kt:(521,543): warning: Viper verification error: Postcondition of pkg_$global$incorrect_exactly_once might not hold. Assertion local$f.special$function_object_call_counter == old(local$f.special$function_object_call_counter) + 1 might not hold. (<no position>)
+
+
+/calls_in_place.kt:(521,543): warning: Function incorrect_exactly_once may not satisfy its contract.
+
+/calls_in_place.kt:(573,627): warning: Wrong invocation kind 'EXACTLY_ONCE' for 'f: (Int) -> Int' specified, the actual invocation kind is 'MORE_THAN_ONCE'.
+
+/calls_in_place.kt:(691,714): info: Generated Viper text for incorrect_at_least_once:
+method pkg_$global$unknown(local$f: Ref) returns (ret: Int)
+  requires dom$Type$isSubtype((dom$TypeOf$typeOf(local$f): dom$Type), dom$Type$Function())
+  requires acc(local$f.special$function_object_call_counter, write)
+  ensures acc(local$f.special$function_object_call_counter, write)
+  ensures old(local$f.special$function_object_call_counter) <=
+    local$f.special$function_object_call_counter
+  ensures true
+
+
+method pkg_$global$incorrect_at_least_once(local$f: Ref) returns (ret: Int)
+  requires dom$Type$isSubtype((dom$TypeOf$typeOf(local$f): dom$Type), dom$Type$Function())
+  requires acc(local$f.special$function_object_call_counter, write)
+  ensures acc(local$f.special$function_object_call_counter, write)
+  ensures old(local$f.special$function_object_call_counter) <=
+    local$f.special$function_object_call_counter
+  ensures local$f.special$function_object_call_counter >
+    old(local$f.special$function_object_call_counter)
+{
+  var anonymous$1: Int
+  anonymous$1 := pkg_$global$unknown(local$f)
+  ret := anonymous$1
+}
+
+/calls_in_place.kt:(691,714): warning: Viper verification error: Postcondition of pkg_$global$incorrect_at_least_once might not hold. Assertion local$f.special$function_object_call_counter > old(local$f.special$function_object_call_counter) might not hold. (<no position>)
+
+
+/calls_in_place.kt:(691,714): warning: Function incorrect_at_least_once may not satisfy its contract.
+
+/calls_in_place.kt:(744,799): warning: Wrong invocation kind 'AT_LEAST_ONCE' for 'f: (Int) -> Int' specified, the actual invocation kind is 'UNKNOWN'.
