+++ plugins/formal-verification/testData/diagnostics/bad_contracts/returns_booleans.fir.diag.txt
@@ -1,0 +1,11 @@
+/returns_booleans.kt:(121,146): info: Generated Viper text for incorrectly_returns_false:
+method pkg_$global$incorrectly_returns_false() returns (ret: Bool)
+  ensures ret == true
+{
+  ret := false
+}
+
+/returns_booleans.kt:(121,146): warning: Viper verification error: Postcondition of pkg_$global$incorrectly_returns_false might not hold. Assertion ret == true might not hold. (<no position>)
+
+
+/returns_booleans.kt:(121,146): warning: Function incorrectly_returns_false may not satisfy its contract.
