+++ plugins/formal-verification/testData/diagnostics/bad_contracts/returns_booleans.fir.txt
@@ -1,0 +1,16 @@
+FILE: returns_booleans.kt
+    @R|kotlin/OptIn|(markerClass = vararg(<getClass>(Q|kotlin/contracts/ExperimentalContracts|))) public final fun incorrectly_returns_false(): R|kotlin/Boolean|
+        [R|Contract description]
+         <
+            Returns(TRUE)
+        >
+     {
+         {
+            R|kotlin/contracts/contract|(<L> = contract@fun R|kotlin/contracts/ContractBuilder|.<anonymous>(): R|kotlin/Unit| <inline=Inline, kind=UNKNOWN>  {
+                this@R|special/anonymous|.R|kotlin/contracts/ContractBuilder.returns|(Boolean(true))
+            }
+            )
+        }
+
+        ^incorrectly_returns_false Boolean(false)
+    }
