+++ plugins/formal-verification/testData/diagnostics/bad_contracts/returns_booleans.kt
@@ -1,0 +1,11 @@
+import kotlin.contracts.contract
+import kotlin.contracts.ExperimentalContracts
+
+@OptIn(ExperimentalContracts::class)
+fun <!FUNCTION_WITH_UNVERIFIED_CONTRACT, VIPER_TEXT, VIPER_VERIFICATION_ERROR!>incorrectly_returns_false<!>(): Boolean {
+    contract {
+        returns(true)
+    }
+    return false
+}
+
