+++ plugins/formal-verification/testData/diagnostics/bad_contracts/returns_null.fir.txt
@@ -1,0 +1,31 @@
+FILE: returns_null.kt
+    @R|kotlin/OptIn|(markerClass = vararg(<getClass>(Q|kotlin/contracts/ExperimentalContracts|))) public final fun returns_null_unverifiable(x: R|kotlin/Int?|): R|kotlin/Int?|
+        [R|Contract description]
+         <
+            Returns(WILDCARD) -> FALSE
+        >
+     {
+         {
+            R|kotlin/contracts/contract|(<L> = contract@fun R|kotlin/contracts/ContractBuilder|.<anonymous>(): R|kotlin/Unit| <inline=Inline, kind=UNKNOWN>  {
+                this@R|special/anonymous|.R|kotlin/contracts/ContractBuilder.returns|().R|kotlin/contracts/SimpleEffect.implies|(Boolean(false))
+            }
+            )
+        }
+
+        ^returns_null_unverifiable Null(null)
+    }
+    @R|kotlin/OptIn|(markerClass = vararg(<getClass>(Q|kotlin/contracts/ExperimentalContracts|))) public final fun non_nullable_returns_null(x: R|kotlin/Int|): R|kotlin/Int|
+        [R|Contract description]
+         <
+            Returns(NULL)
+        >
+     {
+         {
+            R|kotlin/contracts/contract|(<L> = contract@fun R|kotlin/contracts/ContractBuilder|.<anonymous>(): R|kotlin/Unit| <inline=Inline, kind=UNKNOWN>  {
+                this@R|special/anonymous|.R|kotlin/contracts/ContractBuilder.returns|(Null(null))
+            }
+            )
+        }
+
+        ^non_nullable_returns_null R|<local>/x|
+    }
