+++ plugins/formal-verification/testData/diagnostics/good_contracts/calls_in_place.kt
@@ -19,14 +19,6 @@ fun <!VIPER_TEXT!>at_most_once<!>(f : (Int) -> Int) : Int{
 }
 
 @OptIn(ExperimentalContracts::class)
-fun <!FUNCTION_WITH_UNVERIFIED_CONTRACT, VIPER_TEXT, VIPER_VERIFICATION_ERROR!>incorrect_at_most_once<!>(f : (Int) -> Int) : Int{
-    <!WRONG_INVOCATION_KIND!>contract {
-        callsInPlace(f, AT_MOST_ONCE)
-    }<!>
-    return f(at_most_once(f))
-}
-
-@OptIn(ExperimentalContracts::class)
 fun <!VIPER_TEXT!>exactly_once<!>(f : (Int) -> Int) : Int{
     contract {
         callsInPlace(f, EXACTLY_ONCE)
@@ -35,27 +27,11 @@ fun <!VIPER_TEXT!>exactly_once<!>(f : (Int) -> Int) : Int{
 }
 
 @OptIn(ExperimentalContracts::class)
-fun <!FUNCTION_WITH_UNVERIFIED_CONTRACT, VIPER_TEXT, VIPER_VERIFICATION_ERROR!>incorrect_exactly_once<!>(f : (Int) -> Int) : Int{
-    <!WRONG_INVOCATION_KIND!>contract {
-        callsInPlace(f, EXACTLY_ONCE)
-    }<!>
-    return f(f(1))
-}
-
-@OptIn(ExperimentalContracts::class)
 fun <!VIPER_TEXT!>at_least_once<!>(f : (Int) -> Int) : Int{
     contract {
         callsInPlace(f, AT_LEAST_ONCE)
     }
     return exactly_once(f)
-}
-
-@OptIn(ExperimentalContracts::class)
-fun <!FUNCTION_WITH_UNVERIFIED_CONTRACT, VIPER_TEXT, VIPER_VERIFICATION_ERROR!>incorrect_at_least_once<!>(f : (Int) -> Int) : Int{
-    <!WRONG_INVOCATION_KIND!>contract {
-        callsInPlace(f, AT_LEAST_ONCE)
-    }<!>
-    return unknown(f)
 }
 
 @OptIn(ExperimentalContracts::class)
