+++ plugins/formal-verification/testData/diagnostics/good_contracts/inlining.fir.txt
