+++ plugins/formal-verification/testData/diagnostics/good_contracts/is_type_contract.fir.txt
@@ -28,32 +28,3 @@ FILE: is_type_contract.kt
         }
 
     }
-    @R|kotlin/OptIn|(markerClass = vararg(<getClass>(Q|kotlin/contracts/ExperimentalContracts|))) public final fun unverifiableTypeCheck(x: R|kotlin/Int?|): R|kotlin/Boolean|
-        [R|Contract description]
-         <
-            Returns(WILDCARD) -> x is kotlin/Unit
-        >
-     {
-         {
-            R|kotlin/contracts/contract|(<L> = contract@fun R|kotlin/contracts/ContractBuilder|.<anonymous>(): R|kotlin/Unit| <inline=Inline, kind=UNKNOWN>  {
-                this@R|special/anonymous|.R|kotlin/contracts/ContractBuilder.returns|().R|kotlin/contracts/SimpleEffect.implies|((R|<local>/x| is R|kotlin/Unit|))
-            }
-            )
-        }
-
-        ^unverifiableTypeCheck (R|<local>/x| is R|kotlin/String|)
-    }
-    @R|kotlin/OptIn|(markerClass = vararg(<getClass>(Q|kotlin/contracts/ExperimentalContracts|))) public final fun nullableNotNonNullable(x: R|kotlin/Int?|): R|kotlin/Unit|
-        [R|Contract description]
-         <
-            Returns(WILDCARD) -> x is kotlin/Int
-        >
-     {
-         {
-            R|kotlin/contracts/contract|(<L> = contract@fun R|kotlin/contracts/ContractBuilder|.<anonymous>(): R|kotlin/Unit| <inline=Inline, kind=UNKNOWN>  {
-                this@R|special/anonymous|.R|kotlin/contracts/ContractBuilder.returns|().R|kotlin/contracts/SimpleEffect.implies|((R|<local>/x| is R|kotlin/Int|))
-            }
-            )
-        }
-
-    }
