+++ plugins/formal-verification/testData/diagnostics/good_contracts/returns_booleans.fir.diag.txt
@@ -14,19 +14,7 @@ method pkg_$global$returns_false() returns (ret: Bool)
   ret := false
 }
 
-/returns_booleans.kt:(418,443): info: Generated Viper text for incorrectly_returns_false:
-method pkg_$global$incorrectly_returns_false() returns (ret: Bool)
-  ensures ret == true
-{
-  ret := false
-}
-
-/returns_booleans.kt:(418,443): warning: Viper verification error: Postcondition of pkg_$global$incorrectly_returns_false might not hold. Assertion ret == true might not hold. (<no position>)
-
-
-/returns_booleans.kt:(418,443): warning: Function incorrectly_returns_false may not satisfy its contract.
-
-/returns_booleans.kt:(561,578): info: Generated Viper text for conditional_basic:
+/returns_booleans.kt:(418,435): info: Generated Viper text for conditional_basic:
 method pkg_$global$conditional_basic(local$b: Bool) returns (ret: Bool)
   ensures ret == true ==> true
   ensures ret == false ==> local$b
@@ -34,7 +22,7 @@ method pkg_$global$conditional_basic(local$b: Bool) returns (ret: Bool)
   ret := true
 }
 
-/returns_booleans.kt:(755,779): info: Generated Viper text for binary_logic_expressions:
+/returns_booleans.kt:(612,636): info: Generated Viper text for binary_logic_expressions:
 method pkg_$global$binary_logic_expressions(local$a: Bool, local$b: Bool)
   returns (ret: Bool)
   ensures ret == false ==> local$b && false
@@ -43,7 +31,7 @@ method pkg_$global$binary_logic_expressions(local$a: Bool, local$b: Bool)
   ret := true
 }
 
-/returns_booleans.kt:(998,1009): info: Generated Viper text for logical_not:
+/returns_booleans.kt:(855,866): info: Generated Viper text for logical_not:
 method pkg_$global$logical_not(local$b: Bool) returns (ret: Bool)
   ensures ret == true ==> !local$b && local$b
   ensures ret == false ==> local$b || !local$b
@@ -51,7 +39,7 @@ method pkg_$global$logical_not(local$b: Bool) returns (ret: Bool)
   ret := false
 }
 
-/returns_booleans.kt:(1195,1218): info: Generated Viper text for call_fun_with_contracts:
+/returns_booleans.kt:(1052,1075): info: Generated Viper text for call_fun_with_contracts:
 method pkg_$global$binary_logic_expressions(local$a: Bool, local$b: Bool)
   returns (ret: Bool)
   ensures ret == false ==> local$b && false
