+++ plugins/formal-verification/testData/diagnostics/good_contracts/simple.kt
