--- plugins/formal-verification/testData/diagnostics/is_type_contract.fir.diag.txt
@@ -1,52 +1,0 @@
-/is_type_contract.kt:(121,129): info: Generated Viper text for isString:
-field pkg_kotlin$class_String$member_length: Int
-
-method pkg_$global$isString(local$x: dom$Nullable[dom$Any])
-  returns (ret: Bool)
-  requires dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$special$Nullable(dom$Type$Any()))
-  ensures ret == true ==>
-    dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$pkg_kotlin$class_String())
-{
-  ret := dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$pkg_kotlin$class_String())
-}
-
-/is_type_contract.kt:(282,299): info: Generated Viper text for subtypeTransitive:
-method pkg_$global$subtypeTransitive(local$x: dom$Unit)
-  returns (ret: dom$Unit)
-  ensures true ==>
-    dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$special$Nullable(dom$Type$Any()))
-{
-}
-
-/is_type_contract.kt:(515,524): warning: Check for instance is always 'true'.
-
-/is_type_contract.kt:(576,597): info: Generated Viper text for unverifiableTypeCheck:
-field pkg_kotlin$class_String$member_length: Int
-
-method pkg_$global$unverifiableTypeCheck(local$x: dom$Nullable[Int])
-  returns (ret: Bool)
-  requires dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$special$Nullable(dom$Type$Int()))
-  ensures true ==>
-    dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$Unit())
-{
-  ret := dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$pkg_kotlin$class_String())
-}
-
-/is_type_contract.kt:(576,597): warning: Viper verification error: Postcondition of pkg_$global$unverifiableTypeCheck might not hold. Assertion true ==> dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$Unit()) might not hold. (<no position>)
-
-
-/is_type_contract.kt:(576,597): warning: Function unverifiableTypeCheck may not satisfy its contract.
-
-/is_type_contract.kt:(744,766): info: Generated Viper text for nullableNotNonNullable:
-method pkg_$global$nullableNotNonNullable(local$x: dom$Nullable[Int])
-  returns (ret: dom$Unit)
-  requires dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$special$Nullable(dom$Type$Int()))
-  ensures true ==>
-    dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$Int())
-{
-}
-
-/is_type_contract.kt:(744,766): warning: Viper verification error: Postcondition of pkg_$global$nullableNotNonNullable might not hold. Assertion true ==> dom$Type$isSubtype((dom$TypeOf$typeOf(local$x): dom$Type), dom$Type$Int()) might not hold. (<no position>)
-
-
-/is_type_contract.kt:(744,766): warning: Function nullableNotNonNullable may not satisfy its contract.
